+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * update links for new home [jcupitt]
 * various doc fixes [janko-m]
 * update docs for 8.7 [jcupitt]
+* add `Vips::get_suffixes` [jcupitt]
 
 ## Version 2.0.13 (2018-8-6)